@@ -1,7 +1,7 @@
 {
   "$schema": "../../node_modules/ng-packagr/package.schema.json",
   "name": "ngx-owl-carousel-o",
-  "version": "2.0.3",
+  "version": "3.0.0",
   "description": "Angular powered owl-carousel",
   "keywords": [
     "angular",
@@ -24,9 +24,10 @@
     "url": "https://github.com/vitalii-andriiovskyi/ngx-owl-carousel-o/issues"
   },
   "peerDependencies": {
-    "@angular/common": "^8.0.0-rc.0 || ^8.0.0 || ^9.0.0-rc.0",
-    "@angular/core": "^8.0.0-rc.0 || ^8.0.0 || ^9.0.0-rc.0",
-    "rxjs": "^6.0.1"
+    "@angular/common": " ^9.0.0-rc.0 || ^9.0.0",
+    "@angular/core": "^9.0.0-rc.0 || ^9.0.0",
+    "rxjs": "^6.0.1",
+    "tslib": "^1.10.0"
   },
   "main": "bundles/ngx-owl-carousel-o.umd.js",
   "module": "fesm5/ngx-owl-carousel-o.js",
@@ -37,8 +38,5 @@
   "fesm2015": "fesm2015/ngx-owl-carousel-o.js",
   "typings": "ngx-owl-carousel-o.d.ts",
   "metadata": "ngx-owl-carousel-o.metadata.json",
-  "sideEffects": false,
-  "dependencies": {
-    "tslib": "^1.9.0"
-  }
+  "sideEffects": false
 }
