@@ -1,9 +1,9 @@
-import { __decorate, __metadata, __spread, __assign, __extends, __param } from 'tslib';
-import { Injectable, isDevMode, ErrorHandler, InjectionToken, PLATFORM_ID, Inject, Optional, Input, Directive, TemplateRef, ContentChildren, QueryList, Output, HostListener, Component, ElementRef, ChangeDetectorRef, EventEmitter, NgZone, Renderer2, Attribute, HostBinding, NgModule } from '@angular/core';
+import { __decorate, __spread, __assign, __extends, __param } from 'tslib';
+import { Injectable, isDevMode, ErrorHandler, InjectionToken, PLATFORM_ID, Inject, Optional, TemplateRef, Input, Directive, EventEmitter, ElementRef, ChangeDetectorRef, ContentChildren, Output, HostListener, Component, NgZone, Renderer2, Attribute, HostBinding, NgModule } from '@angular/core';
 import { isPlatformBrowser, LocationStrategy, CommonModule } from '@angular/common';
 import { Subject, merge, of, from } from 'rxjs';
 import { EventManager } from '@angular/platform-browser';
-import { tap, filter, switchMap, first, skip, take, map, toArray, delay } from 'rxjs/operators';
+import { tap, filter, switchMap, first, take, skip, map, toArray, delay } from 'rxjs/operators';
 import { ActivatedRoute, Router, NavigationEnd } from '@angular/router';
 import { trigger, state, style, transition, animate } from '@angular/animations';
 
@@ -39,9 +39,11 @@ var ResizeService = /** @class */ (function () {
     ResizeService.prototype.onLoaded = function (event) {
         this.windowWidth = event.target;
     };
+    ResizeService.ctorParameters = function () { return [
+        { type: EventManager }
+    ]; };
     ResizeService = __decorate([
-        Injectable(),
-        __metadata("design:paramtypes", [EventManager])
+        Injectable()
     ], ResizeService);
     return ResizeService;
 }());
@@ -52,6 +54,7 @@ var ResizeService = /** @class */ (function () {
 var OwlCarouselOConfig = /** @class */ (function () {
     function OwlCarouselOConfig() {
         this.items = 3;
+        this.skip_validateItems = false;
         this.loop = false;
         this.center = false;
         this.rewind = false;
@@ -107,6 +110,7 @@ var OwlCarouselOConfig = /** @class */ (function () {
 var OwlOptionsMockedTypes = /** @class */ (function () {
     function OwlOptionsMockedTypes() {
         this.items = 'number';
+        this.skip_validateItems = 'boolean';
         this.loop = 'boolean';
         this.center = 'boolean';
         this.rewind = 'boolean';
@@ -178,13 +182,23 @@ var OwlLogger = /** @class */ (function () {
         }
         console.warn.apply(console, __spread([value], rest));
     };
+    OwlLogger.ctorParameters = function () { return [
+        { type: ErrorHandler }
+    ]; };
     OwlLogger = __decorate([
-        Injectable(),
-        __metadata("design:paramtypes", [ErrorHandler])
+        Injectable()
     ], OwlLogger);
     return OwlLogger;
 }());
 
+/**
+ * Current state information and their tags.
+ */
+var States = /** @class */ (function () {
+    function States() {
+    }
+    return States;
+}());
 /**
  * Enumeration for types.
  * @enum {String}
@@ -194,6 +208,7 @@ var Type;
     Type["Event"] = "event";
     Type["State"] = "state";
 })(Type || (Type = {}));
+;
 /**
  * Enumeration for width.
  * @enum {String}
@@ -204,6 +219,23 @@ var Width;
     Width["Inner"] = "inner";
     Width["Outer"] = "outer";
 })(Width || (Width = {}));
+;
+/**
+ * Model for coords of .owl-stage
+ */
+var Coords = /** @class */ (function () {
+    function Coords() {
+    }
+    return Coords;
+}());
+/**
+ * Model for all current data of carousel
+ */
+var CarouselCurrentData = /** @class */ (function () {
+    function CarouselCurrentData() {
+    }
+    return CarouselCurrentData;
+}());
 var CarouselService = /** @class */ (function () {
     function CarouselService(logger) {
         var _this = this;
@@ -650,7 +682,7 @@ var CarouselService = /** @class */ (function () {
     CarouselService.prototype.setOptions = function (options) {
         var configOptions = new OwlCarouselOConfig();
         var checkedOptions = this._validateOptions(options, configOptions);
-        this._options = __assign({}, configOptions, checkedOptions);
+        this._options = __assign(__assign({}, configOptions), checkedOptions);
     };
     /**
      * Checks whether user's option are set properly. Cheking is based on typings;
@@ -675,7 +707,7 @@ var CarouselService = /** @class */ (function () {
                 if (mockedTypes[key] === 'number') {
                     if (this_1._isNumeric(checkedOptions[key])) {
                         checkedOptions[key] = +checkedOptions[key];
-                        checkedOptions[key] = key === 'items' ? this_1._validateItems(checkedOptions[key]) : checkedOptions[key];
+                        checkedOptions[key] = key === 'items' ? this_1._validateItems(checkedOptions[key], checkedOptions.skip_validateItems) : checkedOptions[key];
                     }
                     else {
                         checkedOptions[key] = setRightOption(mockedTypes[key], key);
@@ -702,6 +734,7 @@ var CarouselService = /** @class */ (function () {
                         if (!isString_1) {
                             checkedOptions[key] = setRightOption(mockedTypes[key], key);
                         }
+                        ;
                     }
                     else {
                         checkedOptions[key] = setRightOption(mockedTypes[key], key);
@@ -716,21 +749,26 @@ var CarouselService = /** @class */ (function () {
         return checkedOptions;
     };
     /**
-     * Checks option items set by user and if it bigger than number of slides then returns number of slides
+     * Checks the option `items` set by user and if it bigger than number of slides, the function returns number of slides
      * @param items option items set by user
+     * @param skip_validateItems option `skip_validateItems` set by user
      * @returns right number of items
      */
-    CarouselService.prototype._validateItems = function (items) {
-        var result;
+    CarouselService.prototype._validateItems = function (items, skip_validateItems) {
+        var result = items;
         if (items > this._items.length) {
-            result = this._items.length;
-            this.logger.log('The option \'items\' in your options is bigger than the number of slides. This option is updated to the current number of slides and the navigation got disabled');
+            if (skip_validateItems) {
+                this.logger.log('The option \'items\' in your options is bigger than the number of slides. The navigation got disabled');
+            }
+            else {
+                result = this._items.length;
+                this.logger.log('The option \'items\' in your options is bigger than the number of slides. This option is updated to the current number of slides and the navigation got disabled');
+            }
         }
         else {
             if (items === this._items.length && (this.settings.dots || this.settings.nav)) {
                 this.logger.log('Option \'items\' in your options is equal to the number of slides. So the navigation got disabled');
             }
-            result = items;
         }
         return result;
     };
@@ -781,7 +819,7 @@ var CarouselService = /** @class */ (function () {
                 }
             }
         }
-        this.settings = __assign({}, this._options, overwrites[match], { items: (overwrites[match] && overwrites[match].items) ? this._validateItems(overwrites[match].items) : this._options.items });
+        this.settings = __assign(__assign(__assign({}, this._options), overwrites[match]), { items: (overwrites[match] && overwrites[match].items) ? this._validateItems(overwrites[match].items, this._options.skip_validateItems) : this._options.items });
         // if (typeof this.settings.stagePadding === 'function') {
         // 	this.settings.stagePadding = this.settings.stagePadding();
         // }
@@ -825,6 +863,7 @@ var CarouselService = /** @class */ (function () {
         this.leave('initializing');
         this._trigger('initialized');
     };
+    ;
     /**
      * Sends all data needed for View
      */
@@ -1053,6 +1092,7 @@ var CarouselService = /** @class */ (function () {
             if (position !== -1) {
                 break;
             }
+            ;
         }
         // }
         if (!this.settings.loop) {
@@ -1092,6 +1132,7 @@ var CarouselService = /** @class */ (function () {
     CarouselService.prototype.is = function (state) {
         return this._states.current[state] && this._states.current[state] > 0;
     };
+    ;
     /**
        * Sets the absolute position of the current item.
        * @param position The new absolute position or nothing to leave it unchanged.
@@ -1130,6 +1171,7 @@ var CarouselService = /** @class */ (function () {
         }
         return Object.keys(this._invalidated);
     };
+    ;
     /**
      * Resets the absolute position of the current item.
      * @param position the absolute position of the new item.
@@ -1530,6 +1572,7 @@ var CarouselService = /** @class */ (function () {
             _this._states.current[stateName]++;
         });
     };
+    ;
     /**
        * Leaves a state.
        * @param name - The state name.
@@ -1542,6 +1585,7 @@ var CarouselService = /** @class */ (function () {
             }
         });
     };
+    ;
     /**
        * Registers an event or state.
        * @param object - The event or state to register.
@@ -1647,9 +1691,11 @@ var CarouselService = /** @class */ (function () {
             y: first.y - second.y
         };
     };
+    CarouselService.ctorParameters = function () { return [
+        { type: OwlLogger }
+    ]; };
     CarouselService = __decorate([
-        Injectable(),
-        __metadata("design:paramtypes", [OwlLogger])
+        Injectable()
     ], CarouselService);
     return CarouselService;
 }());
@@ -1804,6 +1850,7 @@ var NavigationService = /** @class */ (function () {
         this.carouselService.navData = this._navData;
         this.carouselService.dotsData = this._dotsData;
     };
+    ;
     /**
      * Updates navigation buttons's and dots's states
      */
@@ -1856,6 +1903,7 @@ var NavigationService = /** @class */ (function () {
         });
         return finalCurrent;
     };
+    ;
     /**
        * Gets the current succesor/predecessor position.
      * @param sussessor position of slide
@@ -1877,6 +1925,7 @@ var NavigationService = /** @class */ (function () {
         }
         return position;
     };
+    ;
     /**
        * Slides to the next item or page.
        * @param speed The time in milliseconds for the transition.
@@ -1884,6 +1933,7 @@ var NavigationService = /** @class */ (function () {
     NavigationService.prototype.next = function (speed) {
         this.carouselService.to(this._getPosition(true), speed);
     };
+    ;
     /**
      * Slides to the previous item or page.
      * @param speed The time in milliseconds for the transition.
@@ -1891,6 +1941,7 @@ var NavigationService = /** @class */ (function () {
     NavigationService.prototype.prev = function (speed) {
         this.carouselService.to(this._getPosition(false), speed);
     };
+    ;
     /**
      * Slides to the specified item or page.
      * @param position - The position of the item or page.
@@ -1907,6 +1958,7 @@ var NavigationService = /** @class */ (function () {
             this.carouselService.to(position, speed);
         }
     };
+    ;
     /**
      * Moves carousel after user's clicking on any dots
      */
@@ -1925,9 +1977,11 @@ var NavigationService = /** @class */ (function () {
         }
         this.carouselService.to(this.carouselService.relative(position), false);
     };
+    NavigationService.ctorParameters = function () { return [
+        { type: CarouselService }
+    ]; };
     NavigationService = __decorate([
-        Injectable(),
-        __metadata("design:paramtypes", [CarouselService])
+        Injectable()
     ], NavigationService);
     return NavigationService;
 }());
@@ -1970,6 +2024,9 @@ var BrowserWindowRef = /** @class */ (function (_super) {
         enumerable: true,
         configurable: true
     });
+    BrowserWindowRef = __decorate([
+        Injectable()
+    ], BrowserWindowRef);
     return BrowserWindowRef;
 }(WindowRef));
 /**
@@ -2045,6 +2102,9 @@ var BrowserDocumentRef = /** @class */ (function (_super) {
         enumerable: true,
         configurable: true
     });
+    BrowserDocumentRef = __decorate([
+        Injectable()
+    ], BrowserDocumentRef);
     return BrowserDocumentRef;
 }(DocumentRef));
 /**
@@ -2143,6 +2203,7 @@ var AutoplayService = /** @class */ (function () {
         this.carouselService.enter('rotating');
         this._setAutoPlayInterval();
     };
+    ;
     /**
        * Gets a new timeout
        * @param timeout - The interval before the next animation starts.
@@ -2162,12 +2223,14 @@ var AutoplayService = /** @class */ (function () {
             _this.carouselService.next(speed || _this.carouselService.settings.autoplaySpeed);
         }, timeout || this.carouselService.settings.autoplayTimeout);
     };
+    ;
     /**
        * Sets autoplay in motion.
        */
     AutoplayService.prototype._setAutoPlayInterval = function (timeout) {
         this._timeout = this._getNextTimeout(timeout);
     };
+    ;
     /**
      * Stops the autoplay.
      */
@@ -2179,6 +2242,7 @@ var AutoplayService = /** @class */ (function () {
         this.winRef.clearTimeout(this._timeout);
         this.carouselService.leave('rotating');
     };
+    ;
     /**
        * Stops the autoplay.
        */
@@ -2188,6 +2252,7 @@ var AutoplayService = /** @class */ (function () {
         }
         this._paused = true;
     };
+    ;
     /**
      * Manages by autoplaying according to data passed by _changedSettingsCarousel$ Obsarvable
      * @param data object with current position of carousel and type of change
@@ -2241,11 +2306,15 @@ var AutoplayService = /** @class */ (function () {
             this._playAfterTranslated();
         }
     };
+    AutoplayService.ctorParameters = function () { return [
+        { type: CarouselService },
+        { type: undefined, decorators: [{ type: Inject, args: [WINDOW,] }] },
+        { type: undefined, decorators: [{ type: Inject, args: [DOCUMENT,] }] }
+    ]; };
     AutoplayService = __decorate([
         Injectable(),
         __param(1, Inject(WINDOW)),
-        __param(2, Inject(DOCUMENT)),
-        __metadata("design:paramtypes", [CarouselService, Object, Object])
+        __param(2, Inject(DOCUMENT))
     ], AutoplayService);
     return AutoplayService;
 }());
@@ -2309,9 +2378,11 @@ var LazyLoadService = /** @class */ (function () {
         }
         this.carouselService.slidesData[position].load = true;
     };
+    LazyLoadService.ctorParameters = function () { return [
+        { type: CarouselService }
+    ]; };
     LazyLoadService = __decorate([
-        Injectable(),
-        __metadata("design:paramtypes", [CarouselService])
+        Injectable()
     ], LazyLoadService);
     return LazyLoadService;
 }());
@@ -2397,6 +2468,7 @@ var AnimateService = /** @class */ (function () {
             });
         }
     };
+    ;
     /**
      * Handles the end of 'animationend' event
      * @param id Id of slides
@@ -2416,9 +2488,12 @@ var AnimateService = /** @class */ (function () {
         });
         this.carouselService.onTransitionEnd();
     };
+    ;
+    AnimateService.ctorParameters = function () { return [
+        { type: CarouselService }
+    ]; };
     AnimateService = __decorate([
-        Injectable(),
-        __metadata("design:paramtypes", [CarouselService])
+        Injectable()
     ], AnimateService);
     return AnimateService;
 }());
@@ -2471,9 +2546,11 @@ var AutoHeightService = /** @class */ (function () {
             slide.heightState = (i >= start && i < end) ? 'full' : 'nulled';
         });
     };
+    AutoHeightService.ctorParameters = function () { return [
+        { type: CarouselService }
+    ]; };
     AutoHeightService = __decorate([
-        Injectable(),
-        __metadata("design:paramtypes", [CarouselService])
+        Injectable()
     ], AutoHeightService);
     return AutoHeightService;
 }());
@@ -2489,6 +2566,7 @@ var HashService = /** @class */ (function () {
                 fragment: of('no route').pipe(take(1))
             };
         }
+        ;
         if (!this.router) {
             this.router = {
                 navigate: function (commands, extras) { return; }
@@ -2540,13 +2618,15 @@ var HashService = /** @class */ (function () {
             _this.rewind(fragment);
         });
     };
+    HashService.ctorParameters = function () { return [
+        { type: CarouselService },
+        { type: ActivatedRoute, decorators: [{ type: Optional }] },
+        { type: Router, decorators: [{ type: Optional }] }
+    ]; };
     HashService = __decorate([
         Injectable(),
         __param(1, Optional()),
-        __param(2, Optional()),
-        __metadata("design:paramtypes", [CarouselService,
-            ActivatedRoute,
-            Router])
+        __param(2, Optional())
     ], HashService);
     return HashService;
 }());
@@ -2586,6 +2666,7 @@ var CarouselSlideDirective = /** @class */ (function () {
         enumerable: true,
         configurable: true
     });
+    ;
     /**
        * Determines if the input is a Number or something that can be coerced to a Number
        * @param - The input to be tested
@@ -2594,30 +2675,26 @@ var CarouselSlideDirective = /** @class */ (function () {
     CarouselSlideDirective.prototype.isNumeric = function (number) {
         return !isNaN(parseFloat(number));
     };
+    CarouselSlideDirective.ctorParameters = function () { return [
+        { type: TemplateRef }
+    ]; };
     __decorate([
-        Input(),
-        __metadata("design:type", Object)
+        Input()
     ], CarouselSlideDirective.prototype, "id", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", Number),
-        __metadata("design:paramtypes", [Number])
+        Input()
     ], CarouselSlideDirective.prototype, "dataMerge", null);
     __decorate([
-        Input(),
-        __metadata("design:type", Object)
+        Input()
     ], CarouselSlideDirective.prototype, "width", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", Object)
+        Input()
     ], CarouselSlideDirective.prototype, "dotContent", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", Object)
+        Input()
     ], CarouselSlideDirective.prototype, "dataHash", void 0);
     CarouselSlideDirective = __decorate([
-        Directive({ selector: 'ng-template[carouselSlide]' }),
-        __metadata("design:paramtypes", [TemplateRef])
+        Directive({ selector: 'ng-template[carouselSlide]' })
     ], CarouselSlideDirective);
     return CarouselSlideDirective;
 }());
@@ -2629,6 +2706,7 @@ var SlidesOutputData = /** @class */ (function () {
     }
     return SlidesOutputData;
 }());
+;
 var CarouselComponent = /** @class */ (function () {
     function CarouselComponent(el, resizeService, carouselService, navigationService, autoplayService, lazyLoadService, animateService, autoHeightService, hashService, logger, changeDetectorRef, docRef) {
         this.el = el;
@@ -2671,6 +2749,7 @@ var CarouselComponent = /** @class */ (function () {
                 break;
         }
     };
+    ;
     CarouselComponent.prototype.ngOnInit = function () {
         this.spyDataStreams();
         this.carouselWindowWidth = this.el.nativeElement.querySelector('.owl-carousel').clientWidth;
@@ -2747,7 +2826,7 @@ var CarouselComponent = /** @class */ (function () {
             var changedPosition = of(value).pipe(filter(function () { return value.property.name === 'position'; }), switchMap(function () { return from(_this.slidesData); }), skip(value.property.value), take(_this.carouselService.settings.items), map(function (slide) {
                 var clonedIdPrefix = _this.carouselService.clonedIdPrefix;
                 var id = slide.id.indexOf(clonedIdPrefix) >= 0 ? slide.id.slice(clonedIdPrefix.length) : slide.id;
-                return __assign({}, slide, { id: id, isActive: true });
+                return __assign(__assign({}, slide), { id: id, isActive: true });
             }), toArray(), map(function (slides) {
                 return {
                     slides: slides,
@@ -2882,39 +2961,43 @@ var CarouselComponent = /** @class */ (function () {
     CarouselComponent.prototype.startPlayTE = function () {
         this.autoplayService.startPlayingTouchEnd();
     };
+    CarouselComponent.ctorParameters = function () { return [
+        { type: ElementRef },
+        { type: ResizeService },
+        { type: CarouselService },
+        { type: NavigationService },
+        { type: AutoplayService },
+        { type: LazyLoadService },
+        { type: AnimateService },
+        { type: AutoHeightService },
+        { type: HashService },
+        { type: OwlLogger },
+        { type: ChangeDetectorRef },
+        { type: undefined, decorators: [{ type: Inject, args: [DOCUMENT,] }] }
+    ]; };
     __decorate([
-        ContentChildren(CarouselSlideDirective),
-        __metadata("design:type", QueryList)
+        ContentChildren(CarouselSlideDirective)
     ], CarouselComponent.prototype, "slides", void 0);
     __decorate([
-        Output(),
-        __metadata("design:type", Object)
+        Output()
     ], CarouselComponent.prototype, "translated", void 0);
     __decorate([
-        Output(),
-        __metadata("design:type", Object)
+        Output()
     ], CarouselComponent.prototype, "dragging", void 0);
     __decorate([
-        Output(),
-        __metadata("design:type", Object)
+        Output()
     ], CarouselComponent.prototype, "change", void 0);
     __decorate([
-        Output(),
-        __metadata("design:type", Object)
+        Output()
     ], CarouselComponent.prototype, "changed", void 0);
     __decorate([
-        Output(),
-        __metadata("design:type", Object)
+        Output()
     ], CarouselComponent.prototype, "initialized", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", Object)
+        Input()
     ], CarouselComponent.prototype, "options", void 0);
     __decorate([
-        HostListener('document:visibilitychange', ['$event']),
-        __metadata("design:type", Function),
-        __metadata("design:paramtypes", [Event]),
-        __metadata("design:returntype", void 0)
+        HostListener('document:visibilitychange', ['$event'])
     ], CarouselComponent.prototype, "onVisibilityChange", null);
     CarouselComponent = __decorate([
         Component({
@@ -2931,29 +3014,9 @@ var CarouselComponent = /** @class */ (function () {
             ],
             styles: [".owl-theme { display: block; }"]
         }),
-        __param(11, Inject(DOCUMENT)),
-        __metadata("design:paramtypes", [ElementRef,
-            ResizeService,
-            CarouselService,
-            NavigationService,
-            AutoplayService,
-            LazyLoadService,
-            AnimateService,
-            AutoHeightService,
-            HashService,
-            OwlLogger,
-            ChangeDetectorRef, Object])
+        __param(11, Inject(DOCUMENT))
     ], CarouselComponent);
     return CarouselComponent;
-}());
-
-/**
- * Data model for managing classes of .owl-stage DOM element
- */
-var StageData = /** @class */ (function () {
-    function StageData() {
-    }
-    return StageData;
 }());
 
 var StageComponent = /** @class */ (function () {
@@ -3125,6 +3188,7 @@ var StageComponent = /** @class */ (function () {
         this._drag.stage.current = stage;
         this._animate(stage.x - this._drag.stage.start.x);
     };
+    ;
     /**
      * Moves .owl-stage left-right
      * @param coordinate coordinate to be set to .owl-stage
@@ -3166,6 +3230,7 @@ var StageComponent = /** @class */ (function () {
         this.listenerMouseUp();
         this.listenerTouchEnd();
     };
+    ;
     /**
        * Prepares data for dragging carousel. It starts after firing `touchstart` and `mousedown` events.
        * @param event - The event arguments.
@@ -3238,47 +3303,36 @@ var StageComponent = /** @class */ (function () {
     StageComponent.prototype.clear = function (id) {
         this.animateService.clear(id);
     };
+    StageComponent.ctorParameters = function () { return [
+        { type: NgZone },
+        { type: ElementRef },
+        { type: Renderer2 },
+        { type: CarouselService },
+        { type: AnimateService }
+    ]; };
     __decorate([
-        Input(),
-        __metadata("design:type", Object)
+        Input()
     ], StageComponent.prototype, "owlDraggable", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", StageData)
+        Input()
     ], StageComponent.prototype, "stageData", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", Array)
+        Input()
     ], StageComponent.prototype, "slidesData", void 0);
     __decorate([
-        HostListener('mousedown', ['$event']),
-        __metadata("design:type", Function),
-        __metadata("design:paramtypes", [Object]),
-        __metadata("design:returntype", void 0)
+        HostListener('mousedown', ['$event'])
     ], StageComponent.prototype, "onMouseDown", null);
     __decorate([
-        HostListener('touchstart', ['$event']),
-        __metadata("design:type", Function),
-        __metadata("design:paramtypes", [Object]),
-        __metadata("design:returntype", void 0)
+        HostListener('touchstart', ['$event'])
     ], StageComponent.prototype, "onTouchStart", null);
     __decorate([
-        HostListener('touchcancel', ['$event']),
-        __metadata("design:type", Function),
-        __metadata("design:paramtypes", [Object]),
-        __metadata("design:returntype", void 0)
+        HostListener('touchcancel', ['$event'])
     ], StageComponent.prototype, "onTouchCancel", null);
     __decorate([
-        HostListener('dragstart'),
-        __metadata("design:type", Function),
-        __metadata("design:paramtypes", []),
-        __metadata("design:returntype", void 0)
+        HostListener('dragstart')
     ], StageComponent.prototype, "onDragStart", null);
     __decorate([
-        HostListener('selectstart'),
-        __metadata("design:type", Function),
-        __metadata("design:paramtypes", []),
-        __metadata("design:returntype", void 0)
+        HostListener('selectstart')
     ], StageComponent.prototype, "onSelectStart", null);
     StageComponent = __decorate([
         Component({
@@ -3298,12 +3352,7 @@ var StageComponent = /** @class */ (function () {
                     ]),
                 ])
             ]
-        }),
-        __metadata("design:paramtypes", [NgZone,
-            ElementRef,
-            Renderer2,
-            CarouselService,
-            AnimateService])
+        })
     ], StageComponent);
     return StageComponent;
 }());
@@ -3368,54 +3417,46 @@ var OwlRouterLinkDirective = /** @class */ (function () {
         enumerable: true,
         configurable: true
     });
+    OwlRouterLinkDirective.ctorParameters = function () { return [
+        { type: Router },
+        { type: ActivatedRoute },
+        { type: String, decorators: [{ type: Attribute, args: ['tabindex',] }] },
+        { type: Renderer2 },
+        { type: ElementRef }
+    ]; };
     __decorate([
-        Input(),
-        __metadata("design:type", Object)
+        Input()
     ], OwlRouterLinkDirective.prototype, "queryParams", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", String)
+        Input()
     ], OwlRouterLinkDirective.prototype, "fragment", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", String)
+        Input()
     ], OwlRouterLinkDirective.prototype, "queryParamsHandling", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", Boolean)
+        Input()
     ], OwlRouterLinkDirective.prototype, "preserveFragment", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", Boolean)
+        Input()
     ], OwlRouterLinkDirective.prototype, "skipLocationChange", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", Boolean)
+        Input()
     ], OwlRouterLinkDirective.prototype, "replaceUrl", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", Object)
+        Input()
     ], OwlRouterLinkDirective.prototype, "stopLink", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", Object),
-        __metadata("design:paramtypes", [Object])
+        Input()
     ], OwlRouterLinkDirective.prototype, "owlRouterLink", null);
     __decorate([
-        Input(),
-        __metadata("design:type", Boolean),
-        __metadata("design:paramtypes", [Boolean])
+        Input()
     ], OwlRouterLinkDirective.prototype, "preserveQueryParams", null);
     __decorate([
-        HostListener('click'),
-        __metadata("design:type", Function),
-        __metadata("design:paramtypes", []),
-        __metadata("design:returntype", Boolean)
+        HostListener('click')
     ], OwlRouterLinkDirective.prototype, "onClick", null);
     OwlRouterLinkDirective = __decorate([
         Directive({ selector: ':not(a)[owlRouterLink]' }),
-        __param(2, Attribute('tabindex')),
-        __metadata("design:paramtypes", [Router, ActivatedRoute, String, Renderer2, ElementRef])
+        __param(2, Attribute('tabindex'))
     ], OwlRouterLinkDirective);
     return OwlRouterLinkDirective;
 }());
@@ -3502,62 +3543,49 @@ var OwlRouterLinkWithHrefDirective = /** @class */ (function () {
         enumerable: true,
         configurable: true
     });
+    OwlRouterLinkWithHrefDirective.ctorParameters = function () { return [
+        { type: Router },
+        { type: ActivatedRoute },
+        { type: LocationStrategy }
+    ]; };
     __decorate([
-        HostBinding('attr.target'), Input(),
-        __metadata("design:type", String)
+        HostBinding('attr.target'), Input()
     ], OwlRouterLinkWithHrefDirective.prototype, "target", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", Object)
+        Input()
     ], OwlRouterLinkWithHrefDirective.prototype, "queryParams", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", String)
+        Input()
     ], OwlRouterLinkWithHrefDirective.prototype, "fragment", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", String)
+        Input()
     ], OwlRouterLinkWithHrefDirective.prototype, "queryParamsHandling", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", Boolean)
+        Input()
     ], OwlRouterLinkWithHrefDirective.prototype, "preserveFragment", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", Boolean)
+        Input()
     ], OwlRouterLinkWithHrefDirective.prototype, "skipLocationChange", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", Boolean)
+        Input()
     ], OwlRouterLinkWithHrefDirective.prototype, "replaceUrl", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", Object)
+        Input()
     ], OwlRouterLinkWithHrefDirective.prototype, "stopLink", void 0);
     __decorate([
-        HostBinding(),
-        __metadata("design:type", String)
+        HostBinding()
     ], OwlRouterLinkWithHrefDirective.prototype, "href", void 0);
     __decorate([
-        Input(),
-        __metadata("design:type", Object),
-        __metadata("design:paramtypes", [Object])
+        Input()
     ], OwlRouterLinkWithHrefDirective.prototype, "owlRouterLink", null);
     __decorate([
-        Input(),
-        __metadata("design:type", Boolean),
-        __metadata("design:paramtypes", [Boolean])
+        Input()
     ], OwlRouterLinkWithHrefDirective.prototype, "preserveQueryParams", null);
     __decorate([
-        HostListener('click', ['$event.button', '$event.ctrlKey', '$event.metaKey', '$event.shiftKey']),
-        __metadata("design:type", Function),
-        __metadata("design:paramtypes", [Number, Boolean, Boolean, Boolean]),
-        __metadata("design:returntype", Boolean)
+        HostListener('click', ['$event.button', '$event.ctrlKey', '$event.metaKey', '$event.shiftKey'])
     ], OwlRouterLinkWithHrefDirective.prototype, "onClick", null);
     OwlRouterLinkWithHrefDirective = __decorate([
-        Directive({ selector: 'a[owlRouterLink]' }),
-        __metadata("design:paramtypes", [Router, ActivatedRoute,
-            LocationStrategy])
+        Directive({ selector: 'a[owlRouterLink]' })
     ], OwlRouterLinkWithHrefDirective);
     return OwlRouterLinkWithHrefDirective;
 }());
@@ -3565,6 +3593,7 @@ function attrBoolValue(s) {
     return s === '' || !!s;
 }
 
+var routes = [];
 var CarouselModule = /** @class */ (function () {
     function CarouselModule() {
     }
@@ -3580,6 +3609,10 @@ var CarouselModule = /** @class */ (function () {
     ], CarouselModule);
     return CarouselModule;
 }());
+
+/**
+ * Generated bundle index. Do not edit.
+ */
 
 export { CarouselComponent, CarouselModule, CarouselSlideDirective, OwlRouterLinkDirective, OwlRouterLinkWithHrefDirective, SlidesOutputData, NavigationService as ɵa, CarouselService as ɵb, OwlLogger as ɵc, AutoplayService as ɵd, WINDOW as ɵe, WindowRef as ɵf, BrowserWindowRef as ɵg, windowFactory as ɵh, browserWindowProvider as ɵi, windowProvider as ɵj, WINDOW_PROVIDERS as ɵk, DOCUMENT as ɵl, DocumentRef as ɵm, BrowserDocumentRef as ɵn, documentFactory as ɵo, browserDocumentProvider as ɵp, documentProvider as ɵq, DOCUMENT_PROVIDERS as ɵr, LazyLoadService as ɵs, AnimateService as ɵt, AutoHeightService as ɵu, HashService as ɵv, ResizeService as ɵw, StageComponent as ɵx };
 //# sourceMappingURL=ngx-owl-carousel-o.js.map
