@@ -182,12 +182,12 @@ export declare class CarouselService {
      * Invalidated parts within the update process.
      */
     private _invalidated;
-    readonly invalidated: any;
+    get invalidated(): any;
     /**
      * Current state information and their tags.
      */
     private _states;
-    readonly states: States;
+    get states(): States;
     /**
      * Ordered list of workers for the update process.
    */
@@ -269,8 +269,9 @@ export declare class CarouselService {
      */
     private _validateOptions;
     /**
-     * Checks option items set by user and if it bigger than number of slides then returns number of slides
+     * Checks the option `items` set by user and if it bigger than number of slides, the function returns number of slides
      * @param items option items set by user
+     * @param skip_validateItems option `skip_validateItems` set by user
      * @returns right number of items
      */
     private _validateItems;
